requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.2.0
django-crispy-forms==1.7.2
django-dbsettings==0.11.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-synchro==0.7
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.5
oauthlib==3.1.0
pkg-resources==0.0.0
pylint==2.4.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
